added logout popup frontend
